Added suggested changes in PR
