+++ lib/cashubrew/web/controllers/mint_controller.ex
@@ -15,7 +15,9 @@ defmodule Cashubrew.Web.MintController do
 
   def keysets(conn, _params) do
     keysets = Nut02.Impl.keysets()
-    json(conn, Nut02.GetKeysetsResponse.from_keysets(keysets))
+    json(conn, Nut02.Serde.GetKeysetsResponse.from_keysets(keysets))
+  rescue
+    e in RuntimeError -> conn |> put_status(:bad_request) |> json(Nut00.Error.new_error(0, e))
   end
 
   def keys(conn, _params) do
